Migrates GPIOWatcher to new plugin layout
