+++ docs/source/code-samples/plugins/pull/io.GPIOWatcher/example.yaml
