+++ pnp/plugins/pull/io/__init__.py
@@ -1,0 +1,16 @@
+"""I/O related plugins."""
+
+import sys
+
+from pnp.plugins.pull import try_import_pull
+
+__IMPORTS__ = [
+    ('io.gpiowatcher', 'GPIOWatcher')
+]
+
+thismodule = sys.modules[__name__]
+for import_path, clazz in __IMPORTS__:
+    module = try_import_pull(import_path, clazz)
+    setattr(thismodule, clazz, module)
+
+__all__ = [clazz for _, clazz in __IMPORTS__]
